+++ client/src/main/java/replicant/ChannelChangeDescriptor.java
@@ -3,13 +3,14 @@ package replicant;
 import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.realityforge.replicant.shared.SharedConstants;
 import replicant.messages.ChannelChange;
 
 final class ChannelChangeDescriptor
 {
   enum Type
   {
-    ADD, REMOVE, UPDATE
+    ADD, REMOVE, UPDATE, DELETE
   }
 
   @Nonnull
@@ -22,7 +23,11 @@ final class ChannelChangeDescriptor
   @Nonnull
   static ChannelChangeDescriptor from( final int schema, @Nonnull final String channelAction )
   {
-    final Type type = '+' == channelAction.charAt( 0 ) ? Type.ADD : Type.REMOVE;
+    final char commandCode = channelAction.charAt( 0 );
+    final Type type =
+      SharedConstants.CHANNEL_ACTION_ADD == commandCode ?
+      Type.ADD :
+      SharedConstants.CHANNEL_ACTION_REMOVE == commandCode ? Type.REMOVE : Type.DELETE;
     final ChannelAddress address = toAddress( schema, channelAction );
     return new ChannelChangeDescriptor( type, address, null );
   }
@@ -31,8 +36,13 @@ final class ChannelChangeDescriptor
   static ChannelChangeDescriptor from( final int schema, @Nonnull final ChannelChange channelChange )
   {
     final String channelAction = channelChange.getChannel();
+    final char commandCode = channelAction.charAt( 0 );
     final Type type =
-      '+' == channelAction.charAt( 0 ) ? Type.ADD : '-' == channelAction.charAt( 0 ) ? Type.REMOVE : Type.UPDATE;
+      SharedConstants.CHANNEL_ACTION_ADD == commandCode ?
+      Type.ADD :
+      SharedConstants.CHANNEL_ACTION_REMOVE == commandCode ?
+      Type.REMOVE :
+      SharedConstants.CHANNEL_ACTION_DELETE == commandCode ? Type.DELETE : Type.UPDATE;
     final ChannelAddress address = toAddress( schema, channelAction );
     return new ChannelChangeDescriptor( type, address, channelChange.getFilter() );
   }